# Performance test for Caffe efficiency measurements