--- a/samples/dnn/squeezenet_halide.cpp
+++ b/samples/dnn/squeezenet_halide.cpp
@@ -1,10 +1,3 @@
-// This file is part of OpenCV project.
-// It is subject to the license terms in the LICENSE file found in the top-level directory
-// of this distribution and at http://opencv.org/license.html.
-//
-// Copyright (C) 2017, Intel Corporation, all rights reserved.
-// Third party copyrights are property of their respective owners.
-
 // Sample of using Halide backend in OpenCV deep learning module.
 // Based on caffe_googlenet.cpp.