<!-- Please use this line to close one or multiple issues when this pullrequest gets merged
You can add another line right under the first one:
resolves #1234
resolves #1235
-->

### This pullrequest changes

* Added performance test for Caffe
* Remove out-date performance statistics from tutorial
* Added `ENABLE_CXX11` flag into OpenCV+Halide build guide
* Wrappers for DNN's backends and targets

Achieved statistics (minimal observed **median** performance metrics)

|          Model |       DNN, C++ (CPU) |    DNN, Halide (CPU) |          Intel-Caffe + MKLDNN (CPU) | clCaffe (GPU)|
|----------------|----------------|----------------|----------------|---|
|      GoogLeNet |        27.39ms |        39.52ms |         **9.31ms** | 18.59ms|
|        AlexNet |        16.86ms |        28.15ms |        **11.95ms** |15.22ms|
|      ResNet-50 |        62.4ms |        104.39ms |        **22.54ms** |57.77ms|
|SqueezeNet v1.1 |         7.46ms |         8.53ms |         **3.01ms** | 5.58ms|
|   Inception-5h |        **29.03ms** |        41.72ms |          ||
| ENet @ 512x256 |        115.76ms |        **58.42ms** |         ||


* CPU: Intel&reg; Core&trade; i7-6700K CPU @ 4.00GHz
* GPU: Intel&reg; HD Graphics 530 (Skylake GT2)
* Disabled hyper-threading and turbo boost technologies
* Fixed CPU freq
  ```bash
  $ cat /proc/cpuinfo | grep "MHz"
  cpu MHz		: 3877.197
  cpu MHz		: 4000.000
  cpu MHz		: 4000.000
  cpu MHz		: 4000.000
  ```

**Merge with extra**: https://github.com/opencv/opencv_extra/pull/361