@dkurt 2017-08-04
@arrybn, may I ask you to reproduce the results of Intel Caffe with MKLDNN engine? Especially for AlexNet, because `caffe time` tool gives 13ms for it.